scientific-skills/umap-learn/scripts/requirements.txt
umap-learn
numpy
matplotlib
pandas
